hw03/requirements.txt
dlib==19.24.2
face-recognition==1.3.0
streamlit==1.32.2
Pillow==10.2.0
numpy==1.26.3